spanet, initial commit
This is the first step towards implementing a component for SpaNET
controllers.

[SpaNET](https://spanet.com.au/) is an Australian company that
manufactures controllers for spa pools (hot tubs). The controllers have
an RJ45 connection for a wifi module which communicates with the
controller via RS232.

The protocol has been reversed engineered by members of the community,
most notably here:
  - https://gist.github.com/devbobo/b82134291d96f384137e01f39a7f2a01
  - https://github.com/BlaT2512/spanet-api/blob/main/spanet.md

This is a uart component that regularly polls the device for its status,
but the current implementation is incomplete, only exposing three of the
properties and not providing any control as of yet.